Use np.int32 to replace np.int for compatible

+++ common/functions.py
@@ -7,7 +7,7 @@ def identity_function(x):
 
 
 def step_function(x):
-    return np.array(x > 0, dtype=np.int)
+    return np.array(x > 0, dtype=np.int32)
 
 
 def sigmoid(x):
